requirements.txt
# Web Framework
flask==3.0.3
flask-cors==4.0.0

# Machine Learning
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# Data visualization (install manually if needed)
# matplotlib==3.9.1
# seaborn==0.13.2

# Environment variables
python-dotenv==1.0.1

# HTTP requests
requests==2.32.3

# Development tools
pytest==8.3.2
black==24.4.2
flake8==7.1.0